Refactor playfield logic into dedicated module
